fix: onSubmit for custom forms (#552)
