+++ packages/codegen-ui-react/lib/__tests__/__utils__/amplify-renderer-generator.ts
@@ -13,10 +13,25 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-import { StudioTemplateRendererFactory, GenericDataSchema, StudioComponent } from '@aws-amplify/codegen-ui';
+import {
+  StudioTemplateRendererFactory,
+  GenericDataSchema,
+  StudioComponent,
+  getGenericFromDataStore,
+  Schema,
+  StudioForm,
+} from '@aws-amplify/codegen-ui';
+import { AmplifyFormRenderer } from '../../amplify-ui-renderers/amplify-form-renderer';
 import { AmplifyRenderer } from '../../amplify-ui-renderers/amplify-renderer';
-import { ReactRenderConfig } from '../../react-render-config';
+import { ModuleKind, ReactRenderConfig, ScriptKind, ScriptTarget } from '../../react-render-config';
 import { loadSchemaFromJSONFile } from './example-schema';
+
+export const defaultCLIRenderConfig: ReactRenderConfig = {
+  module: ModuleKind.ES2020,
+  target: ScriptTarget.ES2020,
+  script: ScriptKind.JSX,
+  renderTypeDeclarations: true,
+};
 
 export const generateWithAmplifyRenderer = (
   jsonSchemaFile: string,
@@ -31,4 +46,22 @@ export const generateWithAmplifyRenderer = (
   return isSampleCodeSnippet
     ? { componentText: renderer.renderSampleCodeSnippet().compText }
     : renderer.renderComponent();
+};
+
+export const generateWithAmplifyFormRenderer = (
+  formJsonFile: string,
+  dataSchemaJsonFile: string | undefined,
+  renderConfig: ReactRenderConfig = defaultCLIRenderConfig,
+): { componentText: string; declaration?: string } => {
+  let dataSchema: GenericDataSchema | undefined;
+  if (dataSchemaJsonFile) {
+    const dataStoreSchema = loadSchemaFromJSONFile<Schema>(dataSchemaJsonFile);
+    dataSchema = getGenericFromDataStore(dataStoreSchema);
+  }
+  const rendererFactory = new StudioTemplateRendererFactory(
+    (component: StudioForm) => new AmplifyFormRenderer(component, dataSchema, renderConfig),
+  );
+
+  const renderer = rendererFactory.buildRenderer(loadSchemaFromJSONFile<StudioForm>(formJsonFile));
+  return renderer.renderComponent();
 };
